% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{calc_percentile_at_week}
\alias{calc_percentile_at_week}
\title{Calculate the percentile waiting at a given week}
\usage{
calc_percentile_at_week(
  wl_shape,
  week,
  wlsize_col = "wlsize",
  time_col = "months_waited_id"
)
}
\arguments{
\item{wl_shape}{a tibble with columns for the number of months waited (where 0 is 0-1 month)
and volume waiting}

\item{week}{numeric; the week to calculate the percentile for}

\item{wlsize_col}{character length 1; the name of the column containing the volume waiting}

\item{time_col}{character length 1; the name of the column containing the number of months waiting}
}
\description{
Calculate the percentile waiting at a given week
}
