% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{calc_performance}
\alias{calc_performance}
\title{Calculates performance by period from a given data set and target stock}
\usage{
calc_performance(incompletes_data, target_bin)
}
\arguments{
\item{incompletes_data}{tibble; requires column headers of "period",
"months_waited_id" and "value", where value is the count of incomplete
pathways by period and the number of months the patients have waited}

\item{target_bin}{integer; the number of months waited where patients that
have waited for greater or equal to that number of months have breached
performance}
}
\value{
A two column tibble containing "period" and "prop" columns, where
  prop is the proportion of people that are on the waiting list in each
  period that have waited for the time less that specified by the target_bin
}
\description{
Calculates performance by period from a given data set and target stock
}
