% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_forecast.R
\name{calculate_t1_value}
\alias{calculate_t1_value}
\title{pass in the rtt table and calculate the t1 value by either projecting a
linear model through the data (if it is significant) or taking a mean}
\usage{
calculate_t1_value(monthly_rtt)
}
\description{
pass in the rtt table and calculate the t1 value by either projecting a
linear model through the data (if it is significant) or taking a mean
}
