% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_charts.R
\name{click_info}
\alias{click_info}
\title{function to return the data behind where the user has clicked}
\usage{
click_info(data, click_x, facet = NULL)
}
\arguments{
\item{data}{the data underpinning the chart that has been clicked}

\item{click_x}{the x location where the click occurred}

\item{facet}{the months_waited_id that the data was selected from when the
plots are faceted}
}
\description{
function to return the data behind where the user has clicked
}
