% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{convert_clock_stops_to_activity}
\alias{convert_clock_stops_to_activity}
\title{Convert clock-stop counts to activity counts}
\usage{
convert_clock_stops_to_activity(clock_stops)
}
\arguments{
\item{clock_stops}{named list; items must be named with the specialty name,
and values equal to clock-stop counts (both treatment and renege clock-stops)}
}
\description{
Convert clock-stop counts to activity counts
}
\examples{
list(
  "Respiratory Medicine" = 504,
  "Oral Surgery" = 772
) |>
convert_clock_stops_to_activity()
}
