% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{filters_displays}
\alias{filters_displays}
\title{function to create standard names so the subsequent aggregation tasks don't
break. This function is focussed on performing analysis for a single
selection within the tool, rather than analysing trusts/specialties as a
batch process}
\usage{
filters_displays(trust_parents, trusts, comm_parents, comms, spec)
}
\description{
function to create standard names so the subsequent aggregation tasks don't
break. This function is focussed on performing analysis for a single
selection within the tool, rather than analysing trusts/specialties as a
batch process
}
