% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_forecast.R
\name{forecast_function}
\alias{forecast_function}
\title{Forecast n period based on a selected method}
\usage{
forecast_function(rtt_table, number_timesteps = 13, method, percent_change)
}
\arguments{
\item{rtt_table}{tibble containing a record for each period, and a value
column}

\item{number_timesteps}{integer; number of time steps to forecast}

\item{method}{string; "tbats" or "linear"}

\item{percent_change}{numeric vector; if method is "linear" the percentage
uplift required by the final time step relative to the extrapolated first
time step (where 1 is a 1% uplift). The number of items determines the
number of columns in the final tibble}
}
\description{
Forecast n period based on a selected method
}
