% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{ignore_unused_imports}
\alias{ignore_unused_imports}
\title{Prevent check warnings
This is receommended here : https://r-pkgs.org/dependencies-in-practice.html#sec-dependencies-in-imports-r-code}
\usage{
ignore_unused_imports()
}
\description{
Prevent check warnings
This is receommended here : https://r-pkgs.org/dependencies-in-practice.html#sec-dependencies-in-imports-r-code
}
