% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTTshiny.R
\name{RTTshiny}
\alias{RTTshiny}
\title{NHSRtt: A golem shiny app for modelling waiting times using a stock and flow method}
\description{
NHSRtt: A golem shiny app for modelling waiting times using a stock and flow method
}
