% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{check_imported_data}
\alias{check_imported_data}
\title{check the data imported into the app}
\usage{
check_imported_data(imported_data)
}
\value{
list with two items; a message describing the outputs of the check,
  and the resulting data tibble (which will be NULL if the checks have
  failed)
}
\description{
check the data imported into the app
}
