% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calibration.R
\name{create_modelling_data}
\alias{create_modelling_data}
\title{Create the calibration dataset from the main data table}
\usage{
create_modelling_data(data, max_months_waited = 12)
}
\arguments{
\item{data}{table of referrals, competes and incompletes (as different
types); data needs the following field names: trust, specialty, period_id,
type, months_waited_id, value}

\item{max_months_waited}{integer; the stock to pool the stocks that have
waited longer into}
}
\description{
Create the calibration dataset from the main data table
}
